+++ gandi/cli/commands/domain.py
@@ -3,7 +3,7 @@
 import click
 
 from gandi.cli.core.cli import cli
-from gandi.cli.core.utils import output_generic, output_contact_info
+from gandi.cli.core.utils import output_contact_info, output_domain
 from gandi.cli.core.params import pass_gandi
 
 
@@ -26,12 +26,13 @@ def list(gandi, limit):
 @pass_gandi
 def info(gandi, resource):
     """Display information about a domain."""
-    output_keys = ['fqdn', 'nameservers', 'services', 'zone_id', 'tags']
+    output_keys = ['fqdn', 'nameservers', 'services', 'zone_id', 'tags',
+                   'created', 'expires', 'updated']
     contact_field = ['owner', 'admin', 'bill', 'tech', 'reseller']
 
     result = gandi.domain.info(resource)
     output_contact_info(gandi, result['contacts'], contact_field, justify=12)
-    output_generic(gandi, result, output_keys, justify=12)
+    output_domain(gandi, result, output_keys, justify=12)
 
     return result
 
@@ -73,6 +74,23 @@ def create(gandi, resource, domain, duration, owner, admin, tech, bill,
 
     result = gandi.domain.create(_domain, duration, owner, admin, tech, bill,
                                  background)
+    if background:
+        gandi.pretty_echo(result)
+
+    return result
+
+
+@cli.command()
+@click.option('--duration', default=1, prompt=True,
+              type=click.IntRange(min=1, max=10),
+              help='Registration period in years, between 1 and 10.')
+@click.option('--bg', '--background', default=False, is_flag=True,
+              help='Run command in background mode (default=False).')
+@click.argument('domain')
+@pass_gandi
+def renew(gandi, domain, duration, background):
+    """Renew a domain."""
+    result = gandi.domain.renew(domain, duration, background)
     if background:
         gandi.pretty_echo(result)